arm64: dts: Transfer display data in burst mode
To save power, command mode DSI panels can accept all the data for a
given frame in one high-speed burst and enter a blanking low-power
period (BLLP) for the rest of the frame's blanking period. The DDIC can
then deal with showing the right image data at the right time while the
DSI bus is in LP mode.

Test: DVT display turns on and works correctly at 60 and 90 Hz
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: celtare21 <celtare21@gmail.com>